+++ src/lib/configs/project/workers.ts
@@ -50,29 +50,77 @@ const workerInfoSchema = {
   required: ["timestamp", "definition"],
 } as const satisfies JSONSchemaType<WorkerInfo>;
 
+export type Deal = WorkerInfo & {
+  dealId: string;
+  dealIdOriginal: string;
+  chainNetwork: ChainNetwork;
+  chainNetworkId: number;
+};
+
+type Deals = Record<string, Deal>;
+
+export type Host = WorkerInfo & {
+  relayId: string;
+  installation_spells: {
+    host_id: string;
+    spell_id: string;
+    worker_id: string;
+  }[];
+};
+
+type Hosts = Record<string, Host>;
+
+export type DealsAndHosts = {
+  deals?: Deals;
+  hosts?: Hosts;
+};
+
 type ConfigV0 = {
   version: 0;
-  deals?: Record<
-    string,
-    WorkerInfo & {
-      dealId: string;
-      dealIdOriginal: string;
-      chainNetwork: ChainNetwork;
-      chainNetworkId: number;
-    }
-  >;
-  hosts?: Record<
-    string,
-    WorkerInfo & {
-      relayId: string;
-      installation_spells: {
-        host_id: string;
-        spell_id: string;
-        worker_id: string;
-      }[];
-    }
-  >;
-};
+} & DealsAndHosts;
+
+const hostSchema: JSONSchemaType<Host> = {
+  ...workerInfoSchema,
+  properties: {
+    ...workerInfoSchema.properties,
+    installation_spells: {
+      type: "array",
+      description: "A list of installation spells",
+      items: {
+        type: "object",
+        properties: {
+          host_id: { type: "string" },
+          spell_id: { type: "string" },
+          worker_id: { type: "string" },
+        },
+        required: ["host_id", "spell_id", "worker_id"],
+      },
+    },
+    relayId: { type: "string" },
+  },
+  required: [...workerInfoSchema.required, "installation_spells", "relayId"],
+} as const;
+
+const dealSchema: JSONSchemaType<Deal> = {
+  ...workerInfoSchema,
+  properties: {
+    ...workerInfoSchema.properties,
+    dealId: { type: "string" },
+    dealIdOriginal: { type: "string" },
+    chainNetwork: {
+      type: "string",
+      enum: CHAIN_NETWORKS,
+    },
+    chainNetworkId: { type: "number" },
+  },
+  required: [
+    ...workerInfoSchema.required,
+    "dealId",
+    "dealIdOriginal",
+    "chainNetwork",
+    "chainNetworkId",
+  ],
+} as const;
 
 const configSchemaV0: JSONSchemaType<ConfigV0> = {
   $id: `${TOP_LEVEL_SCHEMA_ID}/${WORKERS_CONFIG_FILE_NAME}`,
@@ -85,25 +133,9 @@ const configSchemaV0: JSONSchemaType<ConfigV0> = {
     deals: {
       type: "object",
       description: "A map of created deals",
-      additionalProperties: {
-        ...workerInfoSchema,
-        properties: {
-          ...workerInfoSchema.properties,
-          dealId: { type: "string" },
-          dealIdOriginal: { type: "string" },
-          chainNetwork: {
-            type: "string",
-            enum: CHAIN_NETWORKS,
-          },
-          chainNetworkId: { type: "number" },
-        },
-        required: [
-          ...workerInfoSchema.required,
-          "dealId",
-          "dealIdOriginal",
-          "chainNetwork",
-          "chainNetworkId",
-        ],
+      additionalProperties: dealSchema,
+      properties: {
+        Worker_deployed_using_deals: dealSchema,
       },
       required: [],
       nullable: true,
@@ -111,37 +143,16 @@ const configSchemaV0: JSONSchemaType<ConfigV0> = {
     hosts: {
       type: "object",
       description: "A map of deployed workers",
-      additionalProperties: {
-        ...workerInfoSchema,
-        properties: {
-          ...workerInfoSchema.properties,
-          installation_spells: {
-            type: "array",
-            description: "A list of installation spells",
-            items: {
-              type: "object",
-              properties: {
-                host_id: { type: "string" },
-                spell_id: { type: "string" },
-                worker_id: { type: "string" },
-              },
-              required: ["host_id", "spell_id", "worker_id"],
-            },
-          },
-          relayId: { type: "string" },
-        },
-        required: [
-          ...workerInfoSchema.required,
-          "installation_spells",
-          "relayId",
-        ],
+      additionalProperties: hostSchema,
+      properties: {
+        Worker_deployed_using_direct_hosting: hostSchema,
       },
       required: [],
       nullable: true,
     },
   },
   required: ["version"],
-};
+} as const;
 
 const migrations: Migrations<Config> = [];
 
@@ -164,11 +175,6 @@ const getDefault: GetDefaultConfig<LatestConfig> = () => {
   };
 };
 
-export const initNewReadonlyWorkersConfig = getReadonlyConfigInitFunction(
-  initConfigOptions,
-  getDefault
-);
-
 export const initNewWorkersConfig = getConfigInitFunction(
   initConfigOptions,
   getDefault
@@ -176,7 +182,5 @@ export const initNewWorkersConfig = getConfigInitFunction(
 
 export const initReadonlyWorkersConfig =
   getReadonlyConfigInitFunction(initConfigOptions);
-
-export const initWorkersConfig = getConfigInitFunction(initConfigOptions);
 
 export const workersSchema: JSONSchemaType<LatestConfig> = configSchemaV0;