+++ src/lib/const.ts
@@ -30,6 +30,8 @@ import type { FluenceEnv } from "./multiaddres.js";
 import { FLUENCE_ENV } from "./setupEnvironment.js";
 import { getIsStringUnion } from "./typeHelpers.js";
 
+export const NODE_JS_MAJOR_VERSION = 18;
+
 export const TS_NODE_RECOMMENDED_VERSION = "10.9.1";
 export const TYPESCRIPT_RECOMMENDED_VERSION = "4.8.4";
 export const RUST_WASM32_WASI_TARGET = "wasm32-wasi";
@@ -99,7 +101,6 @@ export const SRC_DIR_NAME = "src";
 export const TS_DIR_NAME = "ts";
 export const JS_DIR_NAME = "js";
 export const TMP_DIR_NAME = "tmp";
-export const TMP_SERVICES_DIR_NAME = "services";
 export const VSCODE_DIR_NAME = ".vscode";
 export const NODE_MODULES_DIR_NAME = "node_modules";
 export const AQUA_DIR_NAME = "aqua";
@@ -109,7 +110,6 @@ export const SPELLS_DIR_NAME = "spells";
 export const NPM_DIR_NAME = "npm";
 export const CARGO_DIR_NAME = "cargo";
 export const BIN_DIR_NAME = "bin";
-export const DOT_BIN_DIR_NAME = ".bin";
 export const COUNTLY_DIR_NAME = "countly";
 
 export const FLUENCE_CONFIG_FILE_NAME = `fluence.${YAML_EXT}`;
@@ -119,15 +119,9 @@ export const USER_SECRETS_CONFIG_FILE_NAME = `user-secrets.${YAML_EXT}`;
 export const CONFIG_FILE_NAME = `config.${YAML_EXT}`;
 export const MODULE_CONFIG_FILE_NAME = `module.${YAML_EXT}`;
 export const SERVICE_CONFIG_FILE_NAME = `service.${YAML_EXT}`;
-export const APP_CONFIG_FILE_NAME = `app.${YAML_EXT}`;
-export const DEPENDENCY_CONFIG_FILE_NAME = `dependency.${YAML_EXT}`;
 export const SPELL_CONFIG_FILE_NAME = `spell.${YAML_EXT}`;
 
-const DEPLOYED_APP_FILE_NAME = "deployed.app";
-
-export const DEPLOYED_APP_AQUA_FILE_NAME = `${DEPLOYED_APP_FILE_NAME}.${AQUA_EXT}`;
 export const DEFAULT_SRC_AQUA_FILE_NAME = `main.${AQUA_EXT}`;
-export const INTERFACES_AQUA_FILE_NAME = `interfaces.${AQUA_EXT}`;
 export const AQUA_SERVICES_FILE_NAME = `services.${AQUA_EXT}`;
 export const AQUA_WORKERS_FILE_NAME = `workers.${AQUA_EXT}`;
 export const SPELL_AQUA_FILE_NAME = `spell.${AQUA_EXT}`;
@@ -138,17 +132,10 @@ export const PACKAGE_JSON_FILE_NAME = `package.${JSON_EXT}`;
 export const TS_CONFIG_FILE_NAME = `tsconfig.${JSON_EXT}`;
 export const EXTENSIONS_JSON_FILE_NAME = `extensions.${JSON_EXT}`;
 export const SETTINGS_JSON_FILE_NAME = `settings.${JSON_EXT}`;
-export const DEPLOY_CONFIG_FILE_NAME = `deploy.${JSON_EXT}`;
-export const APP_SERVICE_JSON_FILE_NAME = `app-service.${JSON_EXT}`;
 
-export const APP_TS_FILE_NAME = `app.${TS_EXT}`;
-export const APP_JS_FILE_NAME = `app.${JS_EXT}`;
-export const DEPLOYED_APP_TS_FILE_NAME = `${DEPLOYED_APP_FILE_NAME}.${TS_EXT}`;
-export const DEPLOYED_APP_JS_FILE_NAME = `${DEPLOYED_APP_FILE_NAME}.${JS_EXT}`;
 export const INDEX_TS_FILE_NAME = `index.${TS_EXT}`;
 export const INDEX_JS_FILE_NAME = `index.${JS_EXT}`;
 
-export const CRATES_TOML = `.crates.${TOML_EXT}`;
 export const CONFIG_TOML = `Config.${TOML_EXT}`;
 export const CARGO_TOML = `Cargo.${TOML_EXT}`;
 
@@ -162,7 +149,7 @@ export const AUTO_GENERATED = "auto-generated";
 export const DEFAULT_DEPLOY_NAME = "default";
 export const DEFAULT_WORKER_NAME = "defaultWorker";
 
-export const KEY_PAIR_FLAG_NAME = "key-pair-name";
+const KEY_PAIR_FLAG_NAME = "key-pair-name";
 export const KEY_PAIR_FLAG = {
   [KEY_PAIR_FLAG_NAME]: Flags.string({
     char: "k",
@@ -176,16 +163,8 @@ export type KeyPairFlag = FromFlagsDef<typeof KEY_PAIR_FLAG>;
 export const NO_INPUT_FLAG_NAME = "no-input";
 export const NO_INPUT_FLAG = {
   [NO_INPUT_FLAG_NAME]: Flags.boolean({
+    default: false,
     description: "Don't interactively ask for any input from the user",
-  }),
-} as const;
-
-export const TIMEOUT_FLAG_NAME = "timeout";
-export const TIMEOUT_FLAG = {
-  [TIMEOUT_FLAG_NAME]: Flags.integer({
-    description: "Timeout used for command execution",
-    helpValue: "<milliseconds>",
-    default: 60000,
   }),
 } as const;
 
@@ -203,6 +182,7 @@ export const NETWORK_FLAG = {
 export const GLOBAL_FLAG_NAME = "global";
 export const GLOBAL_FLAG = {
   [GLOBAL_FLAG_NAME]: Flags.boolean({
+    default: false,
     aliases: ["g"],
     description: `Will override dependencies in a global user's ${CONFIG_FILE_NAME} instead of project's ${FLUENCE_CONFIG_FILE_NAME}`,
   }),
@@ -217,6 +197,7 @@ export const PRIV_KEY_FLAG = {
 
 export const OFF_AQUA_LOGS_FLAG = {
   "off-aqua-logs": Flags.boolean({
+    default: false,
     description:
       "Turns off logs from Console.print in aqua and from IPFS service",
   }),
@@ -224,6 +205,7 @@ export const OFF_AQUA_LOGS_FLAG = {
 
 export const USE_F64_FLAG = {
   f64: Flags.boolean({
+    default: false,
     description:
       "Convert all numbers to f64. Useful for arrays objects that contain numbers of different types in them. Without this flag, numbers will be converted to u64, i64 or f64 depending on their value",
   }),
@@ -231,11 +213,10 @@ export const USE_F64_FLAG = {
 
 export const NO_BUILD_FLAG = {
   "no-build": Flags.boolean({
+    default: false,
     description: "Don't build the project before running the command",
   }),
 };
-
-export type OffAquaLogsFlag = FromFlagsDef<typeof OFF_AQUA_LOGS_FLAG>;
 
 export const IMPORT_FLAG = {
   import: Flags.string({
@@ -243,6 +224,13 @@ export const IMPORT_FLAG = {
       "Path to a directory to import aqua files from. May be used several times",
     helpValue: "<path>",
     multiple: true,
+  }),
+};
+
+export const TRACING_FLAG = {
+  tracing: Flags.boolean({
+    description: "Compile aqua in tracing mode (for debugging purposes)",
+    default: false,
   }),
 };
 
@@ -266,6 +254,7 @@ export const FLUENCE_CLIENT_FLAGS = {
     helpValue: "<milliseconds>",
   }),
   "particle-id": Flags.boolean({
+    default: false,
     description: "Print particle ids when running Fluence js-client",
   }),
 } as const;
@@ -334,10 +323,9 @@ export const MREPL_CARGO_DEPENDENCY = "mrepl";
 export const MARINE_RS_SDK_CARGO_DEPENDENCY = "marine-rs-sdk";
 export const MARINE_RS_SDK_TEST_CARGO_DEPENDENCY = "marine-rs-sdk-test";
 
-export const AQUA_NPM_DEPENDENCY = "@fluencelabs/aqua";
-export const AQUA_LIB_NPM_DEPENDENCY = "@fluencelabs/aqua-lib";
-export const REGISTRY_NPM_DEPENDENCY = "@fluencelabs/registry";
-export const SPELL_NPM_DEPENDENCY = "@fluencelabs/spell";
+const AQUA_LIB_NPM_DEPENDENCY = "@fluencelabs/aqua-lib";
+const REGISTRY_NPM_DEPENDENCY = "@fluencelabs/registry";
+const SPELL_NPM_DEPENDENCY = "@fluencelabs/spell";
 export const JS_CLIENT_NODE_NPM_DEPENDENCY = "@fluencelabs/js-client.node";
 export const JS_CLIENT_API_NPM_DEPENDENCY = "@fluencelabs/js-client.api";
 export const FLUENCE_NETWORK_ENVIRONMENT_NPM_DEPENDENCY =
@@ -345,7 +333,6 @@ export const FLUENCE_NETWORK_ENVIRONMENT_NPM_DEPENDENCY =
 
 export const fluenceNPMDependencies = [
   AQUA_LIB_NPM_DEPENDENCY,
-  AQUA_NPM_DEPENDENCY,
   REGISTRY_NPM_DEPENDENCY,
   SPELL_NPM_DEPENDENCY,
 ] as const;
@@ -365,31 +352,31 @@ export const SEPARATOR = `\n\n${color.yellow(
   `^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^`
 )}\n\n`;
 
-export const MAIN_AQUA_FILE_STATUS_TEXT = `export status
-
--- example of running a service deployed using 'fluence deal deploy'
+const MAIN_AQUA_FILE_STATUS_TEXT = `
+-- example of running services deployed using 'fluence deal deploy'
 -- with worker '${DEFAULT_WORKER_NAME}' which has service 'MyService' with method 'greeting'
 
-func status():
-    workersInfo <- getWorkersInfo()
-    dealId = workersInfo.deals.${DEFAULT_WORKER_NAME}.dealId
-    print = (answer: string, peer: string):
-      Console.print([answer, peer])
+export runDeployedServices
 
-    answers: *string
+data Answer:
+    answer: string
+    peer: string
+
+func runDeployedServices() -> *Answer:
+    workersInfo <- getWorkersInfo()
+    dealId = workersInfo.deals.defaultWorker!.dealId
+    answers: *Answer
     workers <- resolveSubnetwork(dealId)
     for w <- workers! par:
         on w.metadata.peer_id via w.metadata.relay_id:
             answer <- MyService.greeting("fluence")
-            answers <<- answer
-            print(answer, w.metadata.peer_id)
+            answers <<- Answer(answer=answer, peer=w.metadata.relay_id!)
 
-    Console.print("getting answers...")
     join answers[workers!.length - 1]
     par Peer.timeout(PARTICLE_TTL / 2, "TIMED OUT")
-    Console.print("done")`;
+    <- answers`;
 
-export const MAIN_AQUA_FILE_STATUS_TEXT_COMMENT = aquaComment(
+const MAIN_AQUA_FILE_STATUS_TEXT_COMMENT = aquaComment(
   MAIN_AQUA_FILE_STATUS_TEXT
 );
 
@@ -397,8 +384,6 @@ export const MAIN_AQUA_FILE_CONTENT = `aqua Main
 
 import "${AQUA_LIB_NPM_DEPENDENCY}/builtin.aqua"
 import "${REGISTRY_NPM_DEPENDENCY}/subnetwork.aqua"
-import Registry from "${REGISTRY_NPM_DEPENDENCY}/registry-service.aqua"
-import "${SPELL_NPM_DEPENDENCY}/spell_service.aqua"
 
 import "${AQUA_WORKERS_FILE_NAME}"
 import "services.aqua"
@@ -451,7 +436,7 @@ func getInfosInParallel(peers: []PeerId) -> []Info:
     <- infos
 `;
 
-export const DISABLE_TS_AND_ES_LINT = `/* eslint-disable */
+const DISABLE_TS_AND_ES_LINT = `/* eslint-disable */
 // @ts-nocheck`;
 
 const NODES_CONST = "nodes";